requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-dotenv
groq
